Seperate block into seperate components
