+++ app/projects/simple-chat-app/page.tsx
@@ -1,14 +1,10 @@
 "use client";
+import { IconListStatic } from "@/app/components/blocks/IconBlocks";
 import {
-  LeftPicRightText,
-  RightPicLeftText,
-  FullText,
-  FullTextHeaders,
   Banner,
-  IconList,
-  IconListStatic,
-  VideoLoopBlock,
-} from "../../components/blocks/PageBlock";
+  FullTextHeaders,
+} from "@/app/components/blocks/TextImageBlocks";
+import { VideoLoopBlock } from "@/app/components/blocks/VideoBlocks";
 import { usePathname } from "next/navigation";
 
 export default function Event() {
