+++ jsrepo.json
@@ -1,0 +1,12 @@
+{
+  "$schema": "https://unpkg.com/jsrepo@2.4.1/schemas/project-config.json",
+  "repos": ["https://reactbits.dev/ts/tailwind"],
+  "includeTests": false,
+  "includeDocs": false,
+  "watermark": true,
+  "formatter": "prettier",
+  "configFiles": {},
+  "paths": {
+    "*": "./app/components"
+  }
+}
