+++ package.json
@@ -13,7 +13,8 @@
     "@react-three/fiber": "^8.17.12",
     "@types/three": "^0.172.0",
     "canvas-confetti": "^1.9.3",
-    "framer-motion": "^11.5.4",
+    "framer-motion": "^12.18.1",
+    "gsap": "^3.13.0",
     "next": "14.2.10",
     "next-themes": "^0.3.0",
     "react": "^18",
